+++ front/lib/actions/mcp_internal_actions/servers/search/rendering.ts
@@ -4,7 +4,7 @@ import type {
   DataSourceNodeListType,
   SearchQueryResourceType,
 } from "@app/lib/actions/mcp_internal_actions/output_schemas";
-import type { ResolvedDataSourceConfiguration } from "@app/lib/actions/mcp_internal_actions/servers/search/utils";
+import type { ResolvedDataSourceConfiguration } from "@app/lib/actions/mcp_internal_actions/servers/utils";
 import type { ConnectorProvider, CoreAPIContentNode } from "@app/types";
 import type { CoreAPISearchNodesResponse, TimeFrame } from "@app/types";
 import { DATA_SOURCE_NODE_ID } from "@app/types";
